Add hidden content roots
These are invisible publicly but can be used to hold content that can be linked into the site using aliases or anything else that is editable but shouldn't have a public URL (e.g. error pages, content "libraries" or site-wide values).

To make these accessible in the navigation upgrade the whole location bar to use custom dialogues.