Page modification status is now calculated from the content state
This means that a do-undo cycle won't add the page to the modified
list, e.g. if you add something and then delete it or change some text
& then change it back then the page won't be flagged as modified.

Put another way if the page looks the same as the published version
then it won't appear in the list of pages available for publish.