Replace UIDs with the idea of 'singleton' pages
This removes the last bit of developer-only content from the database.

Instead of relying on db-held UID values to retrieve single pages in a reliable way, developers now define singleton types in code and are able to reference their instances using a set of given 'labels'.

Once declared as a singleton editors (including developers) are prevented from creating more than one instance of that type.