Enable page types to render the output of another page as their own
Calling Page::renders with a block results in the page instances rendering the value returned from the block into their template path.

This combines with a separate #render and #render_inline pathway which enables the renderer to ensure that all content within a page is rendered as page content rather than a page.